requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.5.1
python-dotenv==1.0.0
sqlalchemy==2.0.23